Adding OLED to keyboard level
I hope i did this right....
