+++ keyboards/themadnoodle/udon13/udon13.h
@@ -1,0 +1,7 @@
+// Copyright 2021 JasonRen(biu)
+// Copyright 2022 Drashna Jael're (@Drashna Jael're)
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+#pragma once
+
+#include "quantum.h"
